LPD-19822 SingleIndexToMultipleIndexImporter is an internal interface, it should not be registered as osgi service
