+++ modules/dxp/apps/portal-search-tuning/portal-search-tuning-rankings-web/src/main/java/com/liferay/portal/search/tuning/rankings/web/internal/index/creation/instance/lifecycle/RankingIndexPortalInstanceLifecycleListener.java
@@ -10,6 +10,8 @@ import com.liferay.portal.kernel.model.Company;
 import com.liferay.portal.kernel.model.CompanyConstants;
 import com.liferay.portal.search.capabilities.SearchCapabilities;
 import com.liferay.portal.search.engine.adapter.SearchEngineAdapter;
+import com.liferay.portal.search.index.IndexNameBuilder;
+import com.liferay.portal.search.query.Queries;
 import com.liferay.portal.search.spi.index.creation.instance.lifecycle.BaseIndexPortalInstanceLifecycleListener;
 import com.liferay.portal.search.tuning.rankings.index.RankingIndexReader;
 import com.liferay.portal.search.tuning.rankings.index.name.RankingIndexName;
@@ -17,6 +19,10 @@ import com.liferay.portal.search.tuning.rankings.index.name.RankingIndexNameBuil
 import com.liferay.portal.search.tuning.rankings.web.internal.index.RankingIndexCreatorUtil;
 import com.liferay.portal.search.tuning.rankings.web.internal.index.importer.SingleIndexToMultipleIndexImporter;
 
+import java.util.Map;
+
+import org.osgi.framework.BundleContext;
+import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
@@ -68,6 +74,25 @@ public class RankingIndexPortalInstanceLifecycleListener
 				company.getCompanyId()));
 	}
 
+	@Activate
+	@Override
+	protected void activate(
+		BundleContext bundleContext, Map<String, Object> properties) {
+
+		super.activate(bundleContext, properties);
+
+		_singleIndexToMultipleIndexImporter =
+			new SingleIndexToMultipleIndexImporter(
+				_indexNameBuilder, _queries, _rankingIndexReader,
+				_searchEngineAdapter);
+	}
+
+	@Reference
+	private IndexNameBuilder _indexNameBuilder;
+
+	@Reference
+	private Queries _queries;
+
 	@Reference
 	private RankingIndexNameBuilder _rankingIndexNameBuilder;
 
@@ -80,7 +105,6 @@ public class RankingIndexPortalInstanceLifecycleListener
 	@Reference
 	private SearchEngineAdapter _searchEngineAdapter;
 
-	@Reference
 	private SingleIndexToMultipleIndexImporter
 		_singleIndexToMultipleIndexImporter;
